Refactor contentful code a bit, add docs.
