+++ client/src/contentful/ContentfulPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { JsonifiedPageEntryFields, PageEntryFields } from './entries';
+import { JsonifiedPageEntryFields, PageFields } from './content-types';
 import { documentToReactComponents } from '@contentful/rich-text-react-renderer';
 import { BLOCKS } from '@contentful/rich-text-types';
 import { Asset } from 'contentful';
@@ -7,12 +7,12 @@ import { withI18n, withI18nProps } from '@lingui/react';
 import { SupportedLocale } from '../i18n-base';
 
 type LocalizedPages = {
-  [P in SupportedLocale]: JsonifiedPageEntryFields|PageEntryFields
+  [P in SupportedLocale]: JsonifiedPageEntryFields|PageFields
 };
 
 const ContentfulPageWithI18n: React.FC<{locales: LocalizedPages} & withI18nProps> = props => {
   const locale = props.i18n.language as SupportedLocale;
-  const page = props.locales[locale] as PageEntryFields;
+  const page = props.locales[locale] as PageFields;
   const result = documentToReactComponents(page.content, {
     renderNode: {
       [BLOCKS.EMBEDDED_ASSET]: node => {
